% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataset.R
\docType{data}
\name{adaLASSO.coeff}
\alias{adaLASSO.coeff}
\title{AdaLASSO coefficient}
\format{
A data frame with 58 rows (probes relevant to silenced genes) and their corresponding adaLASSO coefficient
\describe{contains coefficient to calculate iMES.}
}
\usage{
adaLASSO.coeff
}
\description{
A data frame storing feature coefficient for model-selected probes
}
\keyword{datasets}
