% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataset.R
\docType{data}
\name{lasso_fea_gene}
\alias{lasso_fea_gene}
\title{LASSO gene features}
\format{
An vector of genes that constitutes iMES
\describe{contains 55 genes.}
}
\usage{
lasso_fea_gene
}
\description{
An vector of genes that constitutes iMES
}
\keyword{datasets}
