% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predRegulon.R
\name{predRegulon}
\alias{predRegulon}
\title{Regulon phenotypes based on genes with epigenetic silencing}
\usage{
predRegulon(
  emat = NULL,
  samples = NULL,
  seed = 20000112,
  fig.path = getwd(),
  fig.name = "heatmap of regulon activity"
)
}
\arguments{
\item{emat}{A numeric transcritomic gene expression with row features (genes) and sample columns and continuous values as input with proper normalisation (e.g., TPM, FPKM, normalized count, or microarray signals).}

\item{samples}{A string value to indicate the samples that will be used to calculate regulon activity; all samples will be used by default.}

\item{seed}{A numeric string to indicate seed for K-mode clustering for reproducibility}

\item{fig.path}{A string value to indicate the output path for storing the regulon activity heatmap.}

\item{fig.name}{A string value to indicate the name of the regulon activity heatmap.}
}
\value{
A DataFrame with rownames of regulons and colnames of samples with input of regulon activity status, and a predictive regulon phenotype based on K-modes clustering (k=2) with a heatmap.
}
\description{
This function inferrs regulon activity of genes that were epigenetically silenced using transcriptomic expression data for patients with clear cell renal cell carcinoma.
}
\author{
Xiaofan Lu
}
