% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataset.R
\docType{data}
\name{rtni_kirc}
\alias{rtni_kirc}
\title{TNI object}
\format{
An object of class TNI
\describe{contains TNI an object of class RTN and can be used for external prediction of regulon activity.}
}
\usage{
rtni_kirc
}
\description{
An R object derived from Transcriptional Network Inference
}
\keyword{datasets}
